requirements.txt
fastapi>=0.118.0,<1.0.0
uvicorn[standard]>=0.37.0,<1.0.0
apscheduler>=3.11.0,<4.0.0
pydantic>=2.11.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0
alembic>=1.16.0,<2.0.0
python-multipart>=0.0.20,<1.0.0
msal>=1.28.0,<2.0.0
itsdangerous>=2.1.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
redis>=5.0.0,<6.0.0
python-magic-bin>=0.4.14,<1.0.0
